requirements.txt
TensorFlow==1.8
grpcio-tools==1.2.1
futures==3.2.0